Refactor controllers and services with interfaces
