+++ src/main/java/bookingManagementSystem/example/bookingManagementSystem/controller/AuthControllerImpl.java
@@ -1,0 +1,30 @@
+package bookingManagementSystem.example.bookingManagementSystem.controller;
+
+import bookingManagementSystem.example.bookingManagementSystem.model.dto.request.LoginRequest;
+import bookingManagementSystem.example.bookingManagementSystem.model.dto.request.RegisterRequest;
+import bookingManagementSystem.example.bookingManagementSystem.model.dto.response.AuthResponse;
+import bookingManagementSystem.example.bookingManagementSystem.service.AuthService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/auth")
+@RequiredArgsConstructor
+public class AuthControllerImpl implements AuthController {
+
+    private final AuthService authService;
+
+    @PostMapping("/register")
+    public ResponseEntity<AuthResponse> register(@RequestBody RegisterRequest request) {
+        return ResponseEntity.ok(authService.register(request));
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<AuthResponse> login(@RequestBody LoginRequest request) {
+        return ResponseEntity.ok(authService.login(request));
+    }
+}
