+++ src/main/java/bookingManagementSystem/example/bookingManagementSystem/controller/UserController.java
@@ -5,37 +5,15 @@ import bookingManagementSystem.example.bookingManagementSystem.model.dto.request
 import bookingManagementSystem.example.bookingManagementSystem.model.dto.response.AuthResponse;
 import bookingManagementSystem.example.bookingManagementSystem.model.dto.response.ResetPasswordResponse;
 import bookingManagementSystem.example.bookingManagementSystem.model.entity.User;
-import bookingManagementSystem.example.bookingManagementSystem.service.UserService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 import java.util.UUID;
 
-@RestController
-@RequestMapping("api/users")
-@RequiredArgsConstructor
-public class UserController {
-    private final UserService userService;
-
-    @GetMapping("/getAll")
-    ResponseEntity<List<User>> getUsers() {
-        List<User> users = userService.getUsers();
-        return ResponseEntity.ok(users);
-    }
-
-    @PostMapping("/register")
-    ResponseEntity<AuthResponse> register(@RequestBody @Valid RegisterRequest request) {
-        return ResponseEntity.ok(userService.register(request));
-    }
-
-    @PutMapping("/resetPassword/{id}")
-    ResponseEntity<ResetPasswordResponse> resetPassword(@PathVariable UUID id, @RequestBody ResetPasswordRequest request){
-        ResetPasswordResponse user = userService.resetPassword(id, request);
-
-        return ResponseEntity.ok(user);
-    }
-
+public interface UserController {
+    ResponseEntity<List<User>> getUsers();
+    ResponseEntity<AuthResponse> register(@RequestBody RegisterRequest request);
+    ResponseEntity<ResetPasswordResponse> resetPassword(@PathVariable UUID id, @RequestBody ResetPasswordRequest request);
 }
