+++ src/test/java/bookingManagementSystem/example/bookingManagementSystem/service/UserServiceTest.java
@@ -7,6 +7,7 @@ import bookingManagementSystem.example.bookingManagementSystem.model.dto.respons
 import bookingManagementSystem.example.bookingManagementSystem.model.entity.User;
 import bookingManagementSystem.example.bookingManagementSystem.repository.UserRepository;
 import bookingManagementSystem.example.bookingManagementSystem.util.JwtUtil;
+import bookingManagementSystem.example.bookingManagementSystem.service.UserServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -41,7 +42,7 @@ class UserServiceTest {
     JwtUtil jwtUtil;
 
     @InjectMocks
-    UserService userService;
+    UserServiceImpl userService;
 
 
     RegisterRequest registerRequest;
